requirements-backup.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
fonttools==4.51.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
packaging==24.0
playwright==1.44.0
psycopg==3.0
psycopg2-binary==2.9.9
pyee==11.1.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
